Cleaning up some of the simpler warnings cluttering the Issue Navigator.
Now that we're building for 64-bit, we have to cast all of our NSIntegers and NSUIntegers before using them in format strings.

+++ Modules/Calendar/MITDayOfTheWeekCell.m
@@ -102,7 +102,7 @@ NSString * const MITPadDayOfTheWeekCellNibName = @"MITPadDayOfTheWeekCell";
 - (void)setDayOfTheMonth:(NSInteger)dayOfTheMonth
 {
     _dayOfTheMonth = dayOfTheMonth;
-    self.dayOfTheMonthLabel.text = [NSString stringWithFormat:@"%d", dayOfTheMonth];
+    self.dayOfTheMonthLabel.text = [NSString stringWithFormat:@"%ld", (long)dayOfTheMonth];
 }
 
 #pragma mark - Lighter Text Color
